requirements.txt
ffmpeg==1.4
gym==0.21.0
Kivy==2.1.0
matplotlib==3.5.2
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
pytest==7.1.2
pywin32==304
stable-baselines3==1.6.0
tensorboard==2.9.1
torch==1.12.0
wandb==0.12.21
ffmpeg==1.4
imageio-ffmpeg==0.4.7
ray[rllib]
pynput
pydantic
vgamepad==0.0.8
pygame==2.1.2
